Align Samples Between Microbiome and Metabolome Data

Prune both omics objects to the linked sample pairs.

Usage

align_samples_between_omics(
  object = NULL,
  microbiome_data = NULL,
  metabolome_data = NULL,
  sample_link = NULL
)

Arguments

object

A microbiome_metabolome_dataset object.

microbiome_data

A microbiome_dataset object. Used when object is not supplied.

metabolome_data

A mass_dataset object. Used when object is not supplied.

sample_link

Optional sample link table used when object is not supplied.

Value

A microbiome_metabolome_dataset object with aligned samples.

Examples

data("global_patterns", package = "microbiomedataset")

metabolome_expression <- as.data.frame(matrix(
  seq_len(4 * ncol(global_patterns@expression_data)),
  nrow = 4,
  ncol = ncol(global_patterns@expression_data)
))
rownames(metabolome_expression) <- paste0("M", seq_len(nrow(metabolome_expression)))
colnames(metabolome_expression) <- global_patterns@sample_info$sample_id

metabolome_data <- massdataset::create_mass_dataset(
  expression_data = metabolome_expression,
  sample_info = global_patterns@sample_info,
  variable_info = data.frame(variable_id = rownames(metabolome_expression))
)

x <- create_microbiome_metabolome_dataset(global_patterns, metabolome_data, align_samples = FALSE)
x <- align_samples_between_omics(x)
nrow(x@sample_link)
#> [1] 26