Prepare Matched Data for Correlation Analysis

Build filtered microbiome and metabolome matrices suitable for correlation analysis.

Usage

prepare_correlation_data(
  object = NULL,
  microbiome_data = NULL,
  metabolome_data = NULL,
  sample_link = NULL,
  microbiome_rank = "Genus",
  microbiome_transform = c("relative", "log", "presence_absence", "clr", "none"),
  metabolome_transform = c("log", "none"),
  min_taxa_prevalence = 0,
  min_taxa_abundance = 0,
  min_metabolite_prevalence = 0,
  min_metabolite_sd = 0
)

Arguments

object

A microbiome_metabolome_dataset object.

microbiome_data

A microbiome_dataset object. Used when object is not supplied.

metabolome_data

A mass_dataset object. Used when object is not supplied.

sample_link

Optional sample link table used when object is not supplied.

microbiome_rank

Optional taxonomy rank used to aggregate microbiome abundance before extraction.

microbiome_transform

Microbiome transformation method.

metabolome_transform

Metabolome transformation method.

min_taxa_prevalence

Minimum number of non-zero samples required for a microbiome feature.

min_taxa_abundance

Minimum mean abundance required for a microbiome feature.

min_metabolite_prevalence

Minimum number of non-zero samples required for a metabolite.

min_metabolite_sd

Minimum standard deviation required for a metabolite.

Value

A list containing filtered matrices and metadata.

Examples

data("demo_crossomics", package = "microbiomedataset")

x <- prepare_correlation_data(
  microbiome_data = demo_crossomics$microbiome_data,
  metabolome_data = demo_crossomics$metabolome_data,
  sample_link = demo_crossomics$sample_link,
  microbiome_rank = "Genus"
)
#> Warning: NaNs produced
dim(x$microbiome_matrix)
#> [1] 60 40