Objects exported from other packages

These objects are imported from other packages. Follow the links below to see their documentation.

dplyr

arrange, arrange, count, count, desc, desc, distinct, filter, filter, glimpse, group_by, group_by, left_join, left_join, mutate, mutate, mutate_all, mutate_at, n, n, pull, pull, relocate, rename, rename, select, select, slice, slice, slice_head, slice_max, slice_min, slice_sample, slice_tail, summarize, transmute, ungroup

magrittr

%>%

massdataset

check_mass_dataset, check_mass_dataset_class, extract_expression_data, extract_process_info, extract_process_info, extract_sample_info, extract_sample_info_note, extract_variable_info, extract_variable_info_note, get_mv_number, get_sample_id, get_sample_number, get_variable_id, get_variable_number, parse_tidymass_parameter, report_parameters

tidyr

nest, unnest